+++ assets/js/broinfo.js
@@ -1,7 +1,7 @@
 {
 function get_navigator_prop (prop, init_val) {
     if (prop in navigator) {
-        return eval("navigator." + prop);
+        return navigator[prop];
     } else {
         return init_val;
     }
@@ -9,7 +9,7 @@ function get_navigator_prop (prop, init_val) {
 
 function get_window_prop (prop, init_val) {
     if (prop in window) {
-        return eval("window." + prop);
+        return window[prop];
     } else {
         return init_val;
     }
@@ -17,7 +17,7 @@ function get_window_prop (prop, init_val) {
 
 function get_screen_prop (prop, init_val) {
     if (prop in window.screen) {
-        return eval("window.screen." + prop);
+        return window.screen[prop];
     } else {
         return init_val;
     }
@@ -25,7 +25,7 @@ function get_screen_prop (prop, init_val) {
 
 function get_document_prop (prop, init_val) {
     if (prop in document) {
-        return eval("document." + prop);
+        return document[prop];
     } else {
         return init_val;
     }
@@ -36,23 +36,11 @@ function is_dark_mode() {
 }
 
 function get_timezone() {
-    // Intl.DateTimeFormat().resolvedOptions().timeZone;
-    if (Intl) {
-        if (Intl.DateTimeFormat) {
-            var v = Intl.DateTimeFormat();
-            if (v) {
-                if (v.resolvedOptions) {
-                    var vv = v.resolvedOptions();
-                    if (vv) {
-                        if (vv.timeZone) {
-                            return vv.timeZone;
-                        }
-                    }
-                }
-            }
-        }
+    try {
+        return Intl.DateTimeFormat().resolvedOptions().timeZone || '';
+    } catch (e) {
+        return '';
     }
-    return '';
 }
 
 const v_oscpu = get_navigator_prop('oscpu', '');
@@ -68,7 +56,7 @@ const v_referrer = get_document_prop('referrer', '');
 const v_screen_width = get_screen_prop('width', null);
 const v_screen_height = get_screen_prop('height', null);
 const v_screen_color_depth = get_screen_prop('colorDepth', null);
-const v_device_pixcel_ratio = get_window_prop('devicePixelRatio', null);
+const v_device_pixel_ratio = get_window_prop('devicePixelRatio', null);
 
 const v_has_local_storage = (typeof localStorage != 'undefined');
 const v_has_session_storage = (typeof sessionStorage != 'undefined');
@@ -91,7 +79,7 @@ return {
         screen_width: v_screen_width,
         screen_height: v_screen_height,
         screen_color_depth: v_screen_color_depth,
-        device_pixcel_ratio: v_device_pixcel_ratio,
+        device_pixel_ratio: v_device_pixel_ratio,
 
         has_local_storage: v_has_local_storage,
         has_session_storage: v_has_session_storage,